netflix-clone-backend/netflix_clone/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.0.0
psycopg2-binary==2.9.7
setuptools==69.0.3